requirements.txt
beautifulsoup4==4.12.2
python-telegram-bot==20.8
python-telegram-bot==13.15
requests==2.31.0
